+++ src/lib/components/forms/form/example/form-fields.ts
@@ -18,14 +18,15 @@ export const select: TerraFormFieldInterface = {
         name: 'Select',
         listBoxValues: [
             {
-                value: 'option1',
-                caption: 'Option 1',
-                position: 0
-            },
-            {
                 value: 'option2',
                 caption: 'Option 2',
-                position: 1
+                position: 2
+            },
+            {
+                value: 'option1',
+                caption: 'Option 1',
+                position: 1,
+                color: 'green'
             }
         ]
     }
@@ -84,7 +85,7 @@ export const listWithChildren: TerraFormFieldInterface = {
     }
 };
 
-export const containerCompontent: TerraFormFieldInterface = {
+export const containerComponent: TerraFormFieldInterface = {
     type: 'portlet',
     options: {
         name: 'Portlet'
@@ -105,6 +106,21 @@ export const containerCompontent: TerraFormFieldInterface = {
 };
 
 export const formFields: TerraKeyValueInterface<TerraFormFieldInterface> = {
+    checkbox: {
+        type: 'checkbox',
+        options: {
+            name: 'checkbox',
+            required: true
+        }
+    },
+    textarea: {
+        type: 'textarea',
+        options: {
+            name: 'Textarea',
+            required: true,
+            maxRows: 2
+        }
+    },
     text: {
         type: 'text',
         options: {
@@ -128,6 +144,14 @@ export const formFields: TerraKeyValueInterface<TerraFormFieldInterface> = {
             minLength: 3
         }
     },
+    date: {
+        type: 'date',
+        defaultValue: '2019-01-01T00:00:00+01:00',
+        options: {
+            name: 'Test date',
+            displayDateFormat: 'DD.MM.YYYY'
+        }
+    },
     listWithChildren: listWithChildren,
-    wrappedContainer: containerCompontent
+    wrappedContainer: containerComponent
 };