+++ src/lib/components/forms/form/form-components/checkbox/checkbox.interface.ts
@@ -1,0 +1,6 @@
+import { TerraFormComponentBaseInterface } from '../terra-form-component-base.interface';
+
+export interface CheckboxInterface extends TerraFormComponentBaseInterface {
+    icon: string;
+    isIndeterminate: boolean;
+}