+++ src/lib/terra-components.module.ts
@@ -18,6 +18,9 @@ import { ModalModule } from 'ngx-bootstrap/modal';
 import { MatIconModule } from '@angular/material/icon';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { A11yModule } from '@angular/cdk/a11y';
+import { MatSelectModule } from '@angular/material/select';
 import { MatInputModule } from '@angular/material/input';
 
 @NgModule({
@@ -39,6 +42,10 @@ import { MatInputModule } from '@angular/material/input';
         MatIconModule,
         MatCheckboxModule,
         MatFormFieldModule,
+        MatDatepickerModule,
+        A11yModule,
+        MatFormFieldModule,
+        MatSelectModule,
         MatInputModule
     ],
     declarations: [...components, ...directives, ...pipes],