+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 11.2.0 (XX.XX.2021)
+
+### Features
+
+-   **terra-form**
+    -   form type **datepicker** replaced deprecated `terra-date-picker` with a `mat-datepicker` variant.
+    -   form type **text**: replaced deprecated `terra-text-input` with a `mat-input` variant.
+    -   form type **checkbox** replaced deprecated `terra-checkbox` with a `mat-checkbox` variant.
+    -   form type **select** replaced deprecated `terra-select-box` with a `mat-select` variant.
+    -   form type **number**: replaced deprecated `terra-number-input` with a `mat-input` variant.
+    -   form type **text-area** replaced deprecated `terra-text-area-input` with a `mat-input` variant.
+
 # 11.1.0 (23.03.2021)
 
 ### Features